requirements.txt
fastapi==0.103.0
psycopg2-binary==2.9.6
uvicorn==0.23.0
pydantic==1.10.2
